feat(controllers): set minimum sync period to prevent hot looping reconciles
kube-dns in GKE sets an upper bound on the TTL to 30 seconds.

For clusters with a large amount of FQDNNetworkPolicy resources, this will
result in very frequent requeues ad infinitum. This is exacerberated by
resources waiting in queue due to the controller only processing
resources non-concurrently.

We set a lower bound to prevent requeuing reconciles unnecessarily.